# package overrides using a mini module system