--- a/pkgs/applications/terminal-emulators/mlterm/default.nix
+++ b/pkgs/applications/terminal-emulators/mlterm/default.nix
@@ -26,35 +26,6 @@
 , libX11
 , libXft
 , cairo
-# List of external tools to create, this default list includes all default
-# tools, as recorded on release 3.9.3.
-, enableTools ? {
-  mlclient = true;
-  mlconfig = true;
-  mlcc = true;
-  mlterm-menu = true;
-  # Note that according to upstream's ./configure script, to disable
-  # mlimgloader you have to disable _all_ tools. See:
-  # https://github.com/arakiken/mlterm/issues/69
-  mlimgloader = true;
-  registobmp = true;
-  mlfc = true;
-}
-# Most of the input methods and other build features are enabled by default,
-# the following attribute set can be used to disable some of them. It's parsed
-# when we set `configureFlags`. If you find other configure Flags that require
-# dependencies, it'd be nice to make that contribution here.
-, enableFeatures ? {
-  uim = !stdenv.isDarwin;
-  ibus = !stdenv.isDarwin;
-  fcitx = !stdenv.isDarwin;
-  m17n = !stdenv.isDarwin;
-  ssh2 = true;
-  bidi = true;
-  # Open Type layout support, (substituting glyphs with opentype fonts)
-  otl = true;
-}
-
 , configuration ? { }
 }:
 
@@ -79,8 +50,11 @@ let
     ];
   };
   inherit (eval) config;
+  # For compat in order to not need to touch the entire drv
   enableGuis = config.gui;
   enableTypeEngines = config.typeEngines;
+  enableTools = config.tools;
+  enableFeatures = config.features;
 in stdenv.mkDerivation (finalAttrs: {
   pname = "mlterm";
   version = "3.9.3";