… pkgs/development/libraries/ffmpeg/options.nix
… pkgs/top-level/all-packages.nix