@AndersonTorres 2024-06-02
Obviouslt my question is: how hard will be to teach this to newcomers?