@Atemu 2024-06-02
Teach what to newcomers? The module system they likely already use in NixOS?