@AndersonTorres 2024-06-02
Yes, given that Nixpkgs is not tied to NixOS.