@AndersonTorres 2024-06-07
> What part do you expect to be hard to teach to newcomers? 

Any part not well documented, to begin with.

https://nixos.org/manual/nixpkgs/stable/#module-system

> Note
> 
> This chapter **is new and not complete yet**. For a gentle introduction to the module system, in the context of NixOS, see [Writing NixOS Modules](https://nixos.org/manual/nixos/unstable/#sec-writing-modules) in the NixOS manual.

(bold mine)