@nixos-discourse 2024-07-21
This pull request has been mentioned on **NixOS Discourse**. There might be relevant details there:

https://discourse.nixos.org/t/a-sad-state-of-nixpkgs/49397/5