@Atemu 2024-09-06
Over the past few weeks, I've been chugging away at an expansion on this idea: making full use of the module system and its type system to define all the core aspects of the derivation (configureFlags and buildInputs).

This new design is quite elegant IMHO; utilising the basic building blocks of the module system to massively cut down the complex and repetitive derivation implementation code into simple data that is even conveniently malleable at "runtime" by end-users.

The line count is roughly double but that is mostly data declaration syntax overhead (`{` and `};`s) and the rest is, as mentioned, mostly data rather than code, so I don't think that's bad at all.  
I much prefer this over the mess that is the function arguments. 

What I especially like is that the packages required for a given feature are now declared in the same place as the feature flag itself. This just feels right to me and also allows for further introspection as you can now programatically figure out which feature pulls in which dependency.

I have not done proper performance testing yet. A quick and dirty test on my laptop showed a significant regression but that was well within the margin of error for a power-limited laptop CPU, so that's hardly conclusive.

This is based on a much older state of ffmpeg, so it's missing a few options that have been added or changed since. I will rebase this in due time but I only care about a working POC for now to fine-tune the design.