@Atemu 2024-09-12
Before:
```
$ hyperfine --warmup 10 -r 30 'nix-instantiate -A ffmpeg-headless'
Benchmark 1: nix-instantiate -A ffmpeg-headless
  Time (mean ± σ):     464.4 ms ±  15.9 ms    [User: 289.1 ms, System: 100.4 ms]
  Range (min … max):   429.5 ms … 486.9 ms    30 runs

$ hyperfine --warmup 10 -r 30 'nix-instantiate -A ffmpeg'
Benchmark 1: nix-instantiate -A ffmpeg
  Time (mean ± σ):     709.7 ms ±  33.3 ms    [User: 450.4 ms, System: 134.3 ms]
  Range (min … max):   649.3 ms … 750.7 ms    30 runs

$ hyperfine --warmup 10 -r 30 'nix-instantiate -A ffmpeg-full'
Benchmark 1: nix-instantiate -A ffmpeg-full
  Time (mean ± σ):     778.4 ms ±  36.8 ms    [User: 472.2 ms, System: 166.5 ms]
  Range (min … max):   710.5 ms … 828.2 ms    30 runs
```
After:
```
$ hyperfine --warmup 10 -r 30 'nix-instantiate -A ffmpeg-headless'
Benchmark 1: nix-instantiate -A ffmpeg-headless
  Time (mean ± σ):     484.4 ms ±  19.0 ms    [User: 316.3 ms, System: 92.1 ms]
  Range (min … max):   441.3 ms … 510.0 ms    30 runs

$ hyperfine --warmup 10 -r 30 'nix-instantiate -A ffmpeg'
Benchmark 1: nix-instantiate -A ffmpeg
  Time (mean ± σ):     727.0 ms ±  30.0 ms    [User: 459.6 ms, System: 148.2 ms]
  Range (min … max):   666.8 ms … 774.0 ms    30 runs

$ hyperfine --warmup 10 -r 30 'nix-instantiate -A ffmpeg-full'
Benchmark 1: nix-instantiate -A ffmpeg-full
  Time (mean ± σ):     815.5 ms ±  38.0 ms    [User: 523.3 ms, System: 157.6 ms]
  Range (min … max):   762.0 ms … 884.0 ms    30 runs
```

There appears to be an effect of about 5% but that only corresponds to ~1σ which isn't really significant, so it's hard to say whether there's an actual negative effect through runtime testing.

Here's the diff of `NIX_SHOW_STATS=1` of `ffmpeg-full` which I expect to show the greatest difference since it contains 3 separate evals of ffmpeg:

```
│  1 │warning: you did not specify '--add-ro↵ │  1 │warning: you did not specify '--add-ro↵
│    │ot'; the result might be removed by th↵ │    │ot'; the result might be removed by th↵
│    │e garbage collector                     │    │e garbage collector
│  2 │/nix/store/3ykq8wg61ibq7xaxcz8h7ygf08i↵ │  2 │/nix/store/4pwzjp1bb4axh7r8ap3jdp777x0
│    │2l8zi-ffmpeg-full-6.1.1.drv!bin         │    │8a4k1-ffmpeg-full-6.1.1.drv!bin
│  3 │{                                       │  3 │{
│  4 │  "cpuTime": 0.46145400404930115,       │  4 │  "cpuTime": 0.5295500159263611,
│  5 │  "envs": {                             │  5 │  "envs": {
│  6 │    "bytes": 15570136,                  │  6 │    "bytes": 27651016,
│  7 │    "elements": 1173009,                │  7 │    "elements": 2054566,
│  8 │    "number": 773258                    │  8 │    "number": 1401811
│  9 │  },                                    │  9 │  },
│ 10 │  "gc": {                               │ 10 │  "gc": {
│ 11 │    "heapSize": 402915328,              │ 11 │    "heapSize": 402915328,
│ 12 │    "totalBytes": 195629904             │ 12 │    "totalBytes": 253786144
│ 13 │  },                                    │ 13 │  },
│ 14 │  "list": {                             │ 14 │  "list": {
│ 15 │    "bytes": 1935720,                   │ 15 │    "bytes": 3137688,
│ 16 │    "concats": 53881,                   │ 16 │    "concats": 78153,
│ 17 │    "elements": 241965                  │ 17 │    "elements": 392211
│ 18 │  },                                    │ 18 │  },
│ 19 │  "nrAvoided": 931140,                  │ 19 │  "nrAvoided": 1577220,
│ 20 │  "nrFunctionCalls": 693286,            │ 20 │  "nrFunctionCalls": 1228708,
│ 21 │  "nrLookups": 349445,                  │ 21 │  "nrLookups": 755979,
│ 22 │  "nrOpUpdateValuesCopied": 4800101,    │ 22 │  "nrOpUpdateValuesCopied": 5003831,
│ 23 │  "nrOpUpdates": 101294,                │ 23 │  "nrOpUpdates": 137413,
│ 24 │  "nrPrimOpCalls": 392853,              │ 24 │  "nrPrimOpCalls": 615520,
│ 25 │  "nrThunks": 1323307,                  │ 25 │  "nrThunks": 2033500,
│ 26 │  "sets": {                             │ 26 │  "sets": {
│ 27 │    "bytes": 96291744,                  │ 27 │    "bytes": 109203312,
│ 28 │    "elements": 5856951,                │ 28 │    "elements": 6502876,
│ 29 │    "number": 161283                    │ 29 │    "number": 322331
│ 30 │  },                                    │ 30 │  },
│ 31 │  "sizes": {                            │ 31 │  "sizes": {
│ 32 │    "Attr": 16,                         │ 32 │    "Attr": 16,

───────────────────────────────────────────────────────────────────────────────────────┐
• 35: warning: you did not specify '--add-root'; the result might be removed by the ga │
───────────────────────────────────────────────────────────────────────────────────────┘
│ 35 │    "Value": 24                         │ 35 │    "Value": 24
│ 36 │  },                                    │ 36 │  },
│ 37 │  "symbols": {                          │ 37 │  "symbols": {
│ 38 │    "bytes": 432883,                    │ 38 │    "bytes": 432598,
│ 39 │    "number": 41660                     │ 39 │    "number": 41663
│ 40 │  },                                    │ 40 │  },
│ 41 │  "values": {                           │ 41 │  "values": {
│ 42 │    "bytes": 47772936,                  │ 42 │    "bytes": 70875288,
│ 43 │    "number": 1990539                   │ 43 │    "number": 2953137
│ 44 │  }                                     │ 44 │  }
│ 45 │}                                       │ 45 │}
```

The only thing I take from this is that there are twice as many function calls now.